+++ packages/playground/src/front/js/utils/js/index.ts
@@ -1,5 +1,7 @@
-import type { languages as Languages } from 'monaco-editor/esm/vs/editor/editor.api.js';
+import type { MonacoNamespace } from '../monaco.js';
 import snippets from './snippets.json';
+
+type Languages = MonacoNamespace['languages'];
 
 interface Range {
   startLineNumber: number;
@@ -8,7 +10,7 @@ interface Range {
   endColumn: number;
 }
 
-function createDependencyProposals(range: Range, languages: typeof Languages) {
+function createDependencyProposals(range: Range, languages: Languages) {
   return Object.values(snippets).map((snippet) => ({
     label: snippet.prefix,
     kind: languages.CompletionItemKind.Function,
@@ -19,7 +21,7 @@ function createDependencyProposals(range: Range, languages: typeof Languages) {
   }));
 }
 
-export function addJsAutocompletion(languages: typeof Languages) {
+export function addJsAutocompletion(languages: Languages) {
   languages.registerCompletionItemProvider('javascript', {
     provideCompletionItems(model, position) {
       const word = model.getWordUntilPosition(position);